+++ protected/humhub/modules/installer/libs/InitialData.php
@@ -424,18 +424,6 @@ class InitialData
         }
 
         $field = new ProfileField();
-        $field->internal_name = "url_googleplus";
-        $field->title = 'Google+ URL';
-        $field->sort_order = 900;
-        $field->profile_field_category_id = $cSocial->id;
-        $field->field_type_class = \humhub\modules\user\models\fieldtype\Text::class;
-        $field->is_system = 1;
-        if ($field->save()) {
-            $field->fieldType->validator = 'url';
-            $field->fieldType->save();
-        }
-
-        $field = new ProfileField();
         $field->internal_name = "url_twitter";
         $field->title = 'Twitter URL';
         $field->sort_order = 1000;