+++ protected/humhub/docs/CHANGELOG.md
@@ -8,6 +8,7 @@ HumHub Change Log
 - Fix #3587: New user tour broken on some template variants
 - Fix #3589: Reduced integrity check memory usage in Like/Comment module
 - Fix #3595: LDAP config file values ignored
+- Fix #3598: Removed Google+ default profile field
 
 
 1.3.14  (June 26, 2019)